Show the navigation in a sidebar on the left
The web version of the jarulfs guide which most people were using, was build with a frameset where the navigation was in a frame on the left side and teh content on the right.
The clear downside of this is that framesets are from the 90th... and they really suck. Especially annoying is that you cannot send someone a link that points to a specific section, because clicking on links does not affect the main url of the page.
The nice part on this though was that the navigation was easily reachable all the time and jumping between section of the guide was much easier than if the navigation is only on the top of the document.

For these reasons I've changed the script lightly to add a div.navigation around the table of contents and changed the styles so that this navigation is always fixed on the left side.
This makes it much more usable imho.